+++ Sources/Extensions/Foundation/FileManagerExtensions.swift
@@ -64,7 +64,7 @@ public extension FileManager {
     ///   - url: URL to save object to.
     ///   - attributes: A dictionary containing the attributes to associate with the new file. You can use these attributes to set the owner and group numbers, file permissions, and modification date. For a list of keys, see FileAttributeKey. Set to nil by default.
     /// - Throws: Throws any errors thrown by JSONEncoder or FileManager.
-    public func encode<T: Encodable>(_ object: T, using encoder: JSONEncoder = JSONEncoder(), to url: URL, attributes: [FileAttributeKey: Any]? = nil) throws -> Bool {
+    public func encodeJSON<T: Encodable>(_ object: T, using encoder: JSONEncoder = JSONEncoder(), to url: URL, attributes: [FileAttributeKey: Any]? = nil) throws -> Bool {
         let data = try encoder.encode(object)
         return createFile(atPath: url.path, contents: data, attributes: attributes)
     }